Move muon_optimizer.py and optimizer.py into optimizers/ subpackage
Group optimizer implementations under trainer_plugins/optimizers/ for
cleaner organization as more optimizers are added. Update relative
imports in optimizer.py (one extra dot, now one level deeper) and the
sibling muon_optimizer import; update the single external import in
base_trainer.py.
